requirements.txt
pillow
vtracer
pyyaml
requests
pymupdf
tqdm
